Factor out object duplication BulkSelectors
